DMP-1574 - enhance /audio/preview and /audio/playback with "Content-Length" and "Content-Range".
