+++ src/main/resources/openapi/audio.yaml
@@ -87,6 +87,11 @@ paths:
           required: true
           description: "Internal identifier for media"
           example: 1
+        - in: header
+          name: range
+          schema:
+            type: string
+          description: "Range header"
       responses:
         200:
           description: OK
@@ -94,7 +99,7 @@ paths:
             audio/mpeg:
               schema:
                 type: string
-                format: binary
+                format: byte
         '400':
           description: Bad Request Error
           content:
